# gatewayservice/openapi.yaml
openapi: 3.0.0
info:
  title: Gatewayservice API
  description: Gateway OpenAPI specification.
  version: 0.2.0
servers:
  - url: http://${{ secrets.DEPLOY_HOST }}:8000
    description: Development server
  - url: http://${{ secrets.DEPLOY_HOST }}:8000
    description: Production server
paths:
  /adduser:
    post:
      summary: Add a new user to the database.
      operationId: addUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: User ID.
                  example: student
                password:
                  type: string
                  description: User password.
                  example: pass
      responses:
        '200':
          description: User added successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  username:
                    type: string
                    description: User ID
                  password:
                    type: string
                    description: Hashed password
                    example: $2b$10$ZKdNYLWFQxzt5Rei/YTc/OsZNi12YiWz30JeUFHNdAt7MyfmkTuvC
                  _id:
                    type: string
                    description: Identification
                    example: 65f756db3fa22d227a4b7c7d
                  createdAt:
                    type: string
                    description: Creation date.
                    example: '2024-03-17T20:47:23.935Z'
                  ___v:
                    type: integer
                    example: '0'                                        
        '400':
          description: Failed to add user.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: getaddrinfo EAI_AGAIN mongodb
  /health:
    get:
      summary: Check the health status of the service.
      operationId: checkHealth
      responses:
        '200':
          description: Service is healthy.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    description: Health status.
                    example: OK
  /login:
    post:
      summary: Log in to the system.
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: User ID.
                  example: student
                password:
                  type: string
                  description: User password.
                  example: pass
      responses:
        '200':
          description: Login successful. Returns user token, username, and creation date.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: User token.
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2NWY3NTZkYjNmYTIyZDIyN2E0YjdjN2QiLCJpYXQiOjE3MTA3MDg3NDUsImV4cCI6MTcxMDcxMjM0NX0.VMG_5DOyQ4GYlJQRcu1I6ICG1IGzuo2Xuei093ONHxw
                  username:
                    type: string
                    description: Username.
                    example: student
                  createdAt:
                    type: string
                    description: Creation date.
                    example: '2024-03-17T20:47:23.935Z'
        '401':
          description: Invalid credentials.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Shows the error info..
                    example: Invalid credentials
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: Internal Server Error
  /generateQuestion:
    get:
      summary: Generate a question.
      operationId: generateQuestion
      responses:
        '200':
          description: Generation successful. Returns the question, the options, the correct option, the image (if neccessary) and the question id
          content:
            application/json:
              schema:
                type: object
                properties:
                  responseQuestion:
                    type: string
                    description: Question.
                    example: ¿Cual es la capital de España?
                  responseOptions:
                    type: array
                    description: Options of the question.
                    example: [Barcelona, Madrid, Oviedo, Valladolid]
                  responseCorrectOption:
                    type: string
                    description: Correct option.
                    example: Madrid
                  responseImage:
                    type: URL
                    description: URL of the image of the question if neccessary.
                    example: https://upload.wikimedia.org/wikipedia/commons/thumb/6/68/Joe_Biden_presidential_portrait.jpg/220px-Joe_Biden_presidential_portrait.jpg
        '400':
          description: Error during the question generator.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Shows the error info.
                    example: numberOfQuestions is not defined
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: Internal Server Error
  /updateQuestion:
    get:
      summary: Update a question.
      operationId: updateQuestion
      parameters:
        time:
          in: query
          description: Time of the question.
          example: 10
          required: true
          schema:
            type: string
        correct:
          in: query
          description: If the question was answered correctly.
          example: true
          required: true
          schema:
            type: boolean
      responses:
        '200':
          description: Update successful. Returns a message ans the updated question
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A message indicating the operation was succesful.
                    example: Tiempo de pregunta actualizado exitosamente
                  updatedQuestion:
                    type: object
                    description: The question updated.
                    example: {enunciado: ¿Cual es la capital de España?, respuesta_correcta: Madrid, respuesta_falsa1: Barcelona, respuesta_falsa2: Oviedo, respuesta_falsa3: Valladolid, 
                      time: 10, correct: true}
        '404':
          description: Cannot find the question.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Shows the error info.
                    example: La pregunta no fue encontrada
        '400':
          description: Error when update the question
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Shows the error info.
                    example: Cannot read property 'time' of undefined
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: Internal Server Error
  /saveGameHistory:
  post:
    summary: Save the game history.
    operationId: saveGameHistory
    responses:
      '200':
        description: Succesful saving the game history
        content:
          application/json:
            schema:
              type: object
              properties:
                success:
                  type: boolean
                  description: Indicates if the operation has been succesful.
                  example: true
      '400':
        description: Error when saving the game history
        content:
          application/json:
            schema:
              type: object
              properties:
                error:
                  type: string
                  description: Shows the error info.
                  example: Error al guardar el historial del juego
      '500':
        description: Internal server error.
        content:
          application/json:
            schema:
              type: object
              properties:
                error:
                  type: string
                  description: Error information.
                  example: Internal Server Error
  /configureGame:
    post:
      summary: Configure the game.
      operationId: configureGame
      responses:
        '200':
          description: Succesful configure the game number of questions
          content:
            application/json:
              schema:
                type: object
                properties:
                  maxQuestions:
                    type: int
                    description: The new number of questions of the question.
                    example: 10
        '400':
          description: Error when configuring the game
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Shows the error info.
                    example: Cannot read property 'valueQuestion' of undefined
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: Internal Server Error
  /gamehistory:
    get:
      summary: Charge the game history.
      operationId: gamehistory
      parameters:
        username:
          in: query
          description: User which game history we want to recover.
          example: Pepito
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Succesful charge the game history
        '400':
          description: Error when charging the game history
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: Internal Server Error
  /getUser:
     get:
      summary: Get an specific user
      operationId: getUser
      parameters:
        username:
          in: query
          description: The name of the user we want to get.
          example: Pepito
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Succesful get the user
          content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: The name of the user.
                  example: Pepito
                email:
                  type: string
                  description: The email of the user.
                  example: pepito@gmail.com
                creado:
                  type: string
                  description: The date where the user account was created.
                  example: 01/01/2024
        '400':
          description: Error when getting the user
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: Internal Server Error
  /getAllUsers:
    get:
      summary: Get all the users.
      operationId: getAllUsers
      responses:
        '200':
          description: Succesful get all the users
          content:
          application/json:
            schema:
              type: object
              properties:
                users:
                  type: array
                  description: The information about the users (username, email, creation date).
                  example: [Pepito, pepito@gmail.com, 01/01/2024]
        '400':
          description: Error when getting the users
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: Internal Server Error
  /getAllQuestions:
    get:
      summary: Get all the questions.
      operationId: getAllQuestions
      responses:
        '200':
          description: Succesful get all the questions
          content:
          application/json:
            schema:
              type: object
              properties:
                users:
                  type: array
                  description: The information about the questions (question an correct answer).
                  example: [¿Cual es la capital de España?,Madrid]
        '400':
          description: Error when getting the questions
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: Internal Server Error
  /topUsers:
    get:
      summary: Charge the top users.
      operationId: topUsers
      responses:
        '200':
          description: Succesful charge the top users
          content:
          application/json:
            schema:
              type: object
              properties:
                users:
                  type: array
                  description: The top 3 users.
                  example: [Pepito, Fulanito, Menganito]
        '400':
          description: Error when charging the top users
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: Internal Server Error
  /endgamestats:
    get:
      summary: Charge the end game statistics.
      operationId: endgamestats
      parameters:
        username:
          in: query
          description: User which game statistics we want to charge.
          example: Pepito
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Succesful charge the end game statitics
          content:
          application/json:
            schema:
              type: object
              properties:
                totalRightQuestions:
                  type: integer
                  description: The number of questions that the user has answered correctly.
                  example: 5
                totalIncorrectQuestions:
                  type: integer
                  description: The number of questions that the user has answered incorrectly.
                  example: 0
                ratio:
                  type: string
                  description: The ratio between the correct and incorrect questions.
                  example: 100%
                totalTime:
                  type: string
                  description: The time the user has spend at the game.
                  example: 20s
                endgameImageWithRatio:
                  type: string
                  description: The end game ratio of questions
                  example: 100%
        '400':
          description: Error when charging the end game statistics
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: Internal Server Error
  /restartGame:
    get:
      summary: Restart the game.
      operationId: endgamestats
      parameters:
        username:
          in: query
          description: User which game statistics we want to charge.
          example: Pepito
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Succesful charge the end game statistics
          content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
                  description: A message indicating the number of questions has been modified.
                  example: Número de preguntas actualizado
                numberOfQuestions:
                  type: integer
                  description: The number of questions that has been modified.
                  example: 5
        '400':
          description: Error when charging the end game statistics
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: Internal Server Error